Change expense menu when a budget has no expenses
When a budget has no expenses, previously the menu would proceed to ask
the user whether they want to add a new expense or if they want to edit
or delete an existing expense. Since there's no existing expense, a
change was made to add some messaging and skip straight to adding a new
expense.